# 8313779: RISC-V: use andn / orn in the MD5 instrinsic