--- a/src/hotspot/cpu/riscv/macroAssembler_riscv.hpp
+++ b/src/hotspot/cpu/riscv/macroAssembler_riscv.hpp
@@ -764,8 +764,8 @@ public:
   void xorrw(Register Rd, Register Rs1, Register Rs2);
 
   // logic with negate
-  void andnr(Register Rd, Register Rs1, Register Rs2);
-  void ornr(Register Rd, Register Rs1, Register Rs2);
+  void andn(Register Rd, Register Rs1, Register Rs2);
+  void orn(Register Rd, Register Rs1, Register Rs2);
 
   // revb
   void revb_h_h(Register Rd, Register Rs, Register tmp = t0);                           // reverse bytes in halfword in lower 16 bits, sign-extend