@gctony 2023-08-04
Was not sure how to differentiate between the `andn` and `orn` in the assembler and the macro assembler. I added an `r` suffix. If there's a better way, let me know!