@gctony 2023-08-07
> I think It should be more consistent if you use 'mv' instead as we did in other places when moving an immediate into register.

Thanks for the suggestion! I thought `mv` was for register to register copy. But I'll change `li` to `mv` in those two places. I'll integrate the change as soon as I retest.