@gctony 2023-08-07
Tests pass (with both `-UseZbb` and `+UseZbb`) with the `li` -> `mv` change. I also confirmed that the generated code is the same.